refractor: add resize icon and enlarge resize area for user specific events
